# docs/docs.go
// Package docs Code generated by swaggo/swag. DO NOT EDIT
package docs

import "github.com/swaggo/swag"

const docTemplate = `{
    "schemes": {{ marshal .Schemes }},
    "swagger": "2.0",
    "info": {
        "description": "{{escape .Description}}",
        "title": "{{.Title}}",
        "contact": {},
        "version": "{{.Version}}"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
        "/add_products": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создаёт новый товар.\n\n**Реализация:**\n\n\u003cpre\u003e\u003ccode\u003efunc handleCreateProduct(c *gin.Context) {\n\tvar p Products\n\tif err := c.ShouldBindJSON(\u0026amp;p); err != nil {\n\t\tc.JSON(http.StatusBadRequest, gin.H{\"error\": err.Error()})\n\t\treturn\n\t}\n\n\tc.JSON(http.StatusCreated, serviceCreate(p))\n}\u003c/code\u003e\u003c/pre\u003e\n",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Post Handler add_products",
                "parameters": [
                    {
                        "description": "Product data",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Products"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created product",
                        "schema": {
                            "$ref": "#/definitions/model.Products"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/token": {
            "post": {
                "description": "Возвращает JWT токен для аутентификации API.\n\n**Реализация:**\n\n\u003cpre\u003e\u003ccode\u003efunc handleAuthToken(c *gin.Context) {\n\ttoken := jwt.NewWithClaims(jwt.SigningMethodHS256, jwt.MapClaims{\n\t\t\"exp\": time.Now().Add(24 * time.Hour).Unix(),\n\t\t\"iat\": time.Now().Unix(),\n\t})\n\tsigned, err := token.SignedString([]byte(settings.JwtSecret))\n\tif err != nil {\n\t\tc.JSON(http.StatusInternalServerError, gin.H{\"error\": err.Error()})\n\t\treturn\n\t}\n\tc.JSON(http.StatusOK, gin.H{\"token\": signed})\n}\u003c/code\u003e\u003c/pre\u003e\n",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Get JWT token",
                "responses": {
                    "200": {
                        "description": "JWT token",
                        "schema": {
                            "$ref": "#/definitions/model.TokenResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/products": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список товаров с фильтрами.\n\n**Реализация:**\n\n\u003cpre\u003e\u003ccode\u003efunc handleGetProducts(c *gin.Context) {\n\tvar filter ProductFilter\n\n\tif s := c.Query(\"min_price\"); s != \"\" {\n\t\tv, err := strconv.Atoi(s)\n\t\tif err != nil {\n\t\t\tc.JSON(http.StatusBadRequest, gin.H{\"error\": \"Некорректный формат min_price\"})\n\t\t\treturn\n\t\t}\n\t\tfilter.MinPrice = \u0026amp;v\n\t}\n\tif s := c.Query(\"max_price\"); s != \"\" {\n\t\tv, err := strconv.Atoi(s)\n\t\tif err != nil {\n\t\t\tc.JSON(http.StatusBadRequest, gin.H{\"error\": \"Некорректный формат max_price\"})\n\t\t\treturn\n\t\t}\n\t\tfilter.MaxPrice = \u0026amp;v\n\t}\n\tif filter.MinPrice != nil \u0026amp;\u0026amp; filter.MaxPrice != nil \u0026amp;\u0026amp; *filter.MinPrice \u0026gt; *filter.MaxPrice {\n\t\tc.JSON(http.StatusBadRequest, gin.H{\"error\": \"min_price должен быть меньше или равен max_price\"})\n\t\treturn\n\t}\n\n\tfilter.InStock = parseInStock(c.Query(\"in_stock\"))\n\n\tc.JSON(http.StatusOK, gin.H{\"data\": serviceGetAll(filter)})\n}\u003c/code\u003e\u003c/pre\u003e\n",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Handler Get Products with filters: min, max, in_stock",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Minimum price filter",
                        "name": "min_price",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum price filter",
                        "name": "max_price",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by stock availability",
                        "name": "in_stock",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of products",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Products"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает один товар\n\n**Реализация:**\n\n\u003cpre\u003e\u003ccode\u003efunc handleGetProductByID(c *gin.Context) {\n\tid, err := strconv.ParseInt(c.Param(\"id\"), 10, 64)\n\tif err != nil {\n\t\tc.JSON(http.StatusBadRequest, gin.H{\"error\": \"Некорректный формат id\"})\n\t\treturn\n\t}\n\n\tp := serviceGetByID(id)\n\tif p == nil {\n\t\tc.JSON(http.StatusNotFound, gin.H{\"error\": \"Продукт не найден\"})\n\t\treturn\n\t}\n\n\tc.JSON(http.StatusOK, gin.H{\"data\": p})\n}\u003c/code\u003e\u003c/pre\u003e\n",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Handler Get Products by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product",
                        "schema": {
                            "$ref": "#/definitions/model.Products"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновляет существующий товар по ID.\n\n**Реализация:**\n\n\u003cpre\u003e\u003ccode\u003efunc handleUpdateProduct(c *gin.Context) {\n\tid, err := strconv.ParseInt(c.Param(\"id\"), 10, 64)\n\tif err != nil {\n\t\tc.JSON(http.StatusBadRequest, gin.H{\"error\": \"Некорректный формат id\"})\n\t\treturn\n\t}\n\n\tvar p Products\n\tif err := c.ShouldBindJSON(\u0026amp;p); err != nil {\n\t\tc.JSON(http.StatusBadRequest, gin.H{\"error\": err.Error()})\n\t\treturn\n\t}\n\n\tupdated := serviceUpdate(id, p)\n\tif updated == nil {\n\t\tc.JSON(http.StatusNotFound, gin.H{\"error\": \"Продукт не найден\"})\n\t\treturn\n\t}\n\n\tc.JSON(http.StatusOK, updated)\n}\u003c/code\u003e\u003c/pre\u003e\n",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Update product by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Product data",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Products"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated product",
                        "schema": {
                            "$ref": "#/definitions/model.Products"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удаляет товар по ID.\n\n**Реализация:**\n\n\u003cpre\u003e\u003ccode\u003efunc handleDeleteProduct(c *gin.Context) {\n\tid, err := strconv.ParseInt(c.Param(\"id\"), 10, 64)\n\tif err != nil {\n\t\tc.JSON(http.StatusBadRequest, gin.H{\"error\": \"Некорректный формат id\"})\n\t\treturn\n\t}\n\n\tif !serviceDelete(id) {\n\t\tc.JSON(http.StatusNotFound, gin.H{\"error\": \"Продукт не найден\"})\n\t\treturn\n\t}\n\n\tc.JSON(http.StatusOK, gin.H{\"message\": fmt.Sprintf(\"Продукт %d удалён\", id)})\n}\u003c/code\u003e\u003c/pre\u003e\n",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Delete product by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deletion confirmation",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "model.Products": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "in_stock": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                }
            }
        },
        "model.TokenResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer {token}\" to authorize.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}`

// SwaggerInfo holds exported Swagger Info so clients can modify it
var SwaggerInfo = &swag.Spec{
	Version:          "1.0",
	Host:             "localhost:8090",
	BasePath:         "",
	Schemes:          []string{},
	Title:            "Products API KR_17_03",
	Description:      "Simple CRUD API for Products with JWT auth.",
	InfoInstanceName: "swagger",
	SwaggerTemplate:  docTemplate,
	LeftDelim:        "{{",
	RightDelim:       "}}",
}

func init() {
	swag.Register(SwaggerInfo.InstanceName(), SwaggerInfo)
}